Create LICENSE file
License is Affero GPL v3. This allows all the normal freedoms to use and copy the source code, but requires that if you create a service using a modified version of this code you must release that code under the same license. This ensures that users of a service derived from this code-base will have the same freedoms as was intended for this project.