Update target framework to .NET Core 2.1
Updated project files and docker files.

The build stage for datadock.web now requires node to be explicitly
installed (it was present in the previous sdk image but has been removed
in 2.1)

Also fixed a problem with the COPY command in the build stage of the
datadock.web Dockerfile that was being rejected by the Docker daemon
because the target directory path did not end with a /. This has
required splitting the copy into separate steps for each directory.