Refactor the DataDockUrlHelper into an injectable service (#107)
Changed the DataDockUrlHelper into IDataDockUriService and added a few more of the common URI patterns to the interface definition. There is a default implementation that implements the standard DataDock URI patterns, but it is now an injectable service which would allow for later support for injecting a more configurable implementation. At the same time, created some basic unit tests for this service.