Legacy data import (#108)
Work to support #54

This basic importer can be built as a docker image that includes the console app and the data to be imported. It can be run via the docker command-line like so:

    docker run --rm -it --network container:dockercompose14270827642735103497_elasticsearch_1 datadock.import:latest /data/20181019_datasets.json /data/20181019_schemas.json http://localhost:9200

NOTE: use the --network option to let the image use the same network stack as the elasticsearch container. This means it can then use the localhost address for the ES endpoint.